requirements.txt
# requests>=2.27.1
# numpy>=1.14.0
# pandas>=0.23.4
# gdown>=3.10.1
# tqdm>=4.30.0
# Pillow>=5.2.0
# opencv-python>=4.5.5.64
# tensorflow>=1.9.0
# keras>=2.2.0
# Flask>=1.1.2
# flask_cors>=4.0.1
# mtcnn>=0.1.0
# retina-face>=0.0.1
# fire>=0.4.0
# gunicorn>=20.1.0
deepface==0.0.93
fastapi
uvicorn
python-multipart
opencv-python>=4.5.5.64
numpy>=1.14.0
Pillow>=5.2.0
tensorflow==2.19.0
tf-keras